Undo renaming of formatDate.js to date.js
It turned out that I do not require an additional date function.
